requirements.txt
fastapi
uvicorn
python-dotenv
crewai
masumi
pydantic
python-multipart
httpx
fredapi